requirements.txt
pyautogui
pillow
pytesseract
googletrans
time
re
sys
requests
